kis változtatások
